+++ packages/calcite-components/src/tests/commonTests/t9n.ts
@@ -3,7 +3,7 @@ import { LitElement, PublicLitElement } from "@arcgis/lumina";
 import { E2EPage, E2EElement } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { expect, it, beforeEach } from "vitest";
 import { MessageBundle } from "../../utils/t9n";
-import { IntrinsicElementsWithProp, newProgrammaticE2EPage } from "../utils";
+import { IntrinsicElementsWithProp, newProgrammaticE2EPage } from "../utils/puppeteer";
 import { getTagAndPage } from "./utils";
 import { ComponentTag, ComponentTestSetup } from "./interfaces";